+++ src/lib/Services/FPMPool.php
@@ -5,20 +5,22 @@
   use \SharedHosting\Utility\Validation;
 
   class FPMPool implements Service {
-    protected $db       = null;
-    protected $files    = [];
-    protected $fwd      = true;
-    protected $username = null;
-    protected $versions = [];
+    protected $db        = null;
+    protected $files     = [];
+    protected $fwd       = true;
+    protected $overwrite = false;
+    protected $username  = null;
+    protected $versions  = [];
 
-    public function __construct(string $username) {
+    public function __construct(string $username, bool $overwrite = false) {
       $this->db = $GLOBALS['db'];
       // Check the provided username for validity
       Validation::username($username);
       // Assign the provided username to an internal property
-      $this->username = $username;
-      $this->versions =  self::fetchVersions();
-      $this->files    = $this->fetchFiles();
+      $this->username  = $username;
+      $this->overwrite = $overwrite;
+      $this->versions  =  self::fetchVersions();
+      $this->files     = $this->fetchFiles();
     }
 
     protected function fetchFiles(): array {
@@ -72,9 +74,11 @@
     public static function flushCommon() {
       // Flush a common include for each PHP version
       foreach (self::fetchVersions() as $version) {
-        file_put_contents('/etc/shared-hosting/php'.$version.'-common.conf',
-          str_replace('{{VERSION}}', $version,
-          file_get_contents(__PROJECTROOT__.'/php-common.cnf')));
+        $file = '/etc/shared-hosting/php'.$version.'-common.conf';
+        // Only write the file if it doesn't exist, or overwrite requested
+        if (!file_exists($file) || $this->overwrite)
+          file_put_contents($file, str_replace('{{VERSION}}', $version,
+            file_get_contents(__PROJECTROOT__.'/php-common.cnf')));
       }
     }