+++ src/lib/Commands/GetDKIM.php
@@ -33,7 +33,10 @@
            ->setHelp        ('This command will print the DKIM DNS record for '.
                              'a given domain name.')
            ->addArgument    ('domain', InputArgument::REQUIRED,
-                             'Any domain associated with a site');;
+                             'Any domain associated with a site')
+           ->addOption      ('get-fqdn', null, InputOption::VALUE_NONE,
+                             'Fetch the Fully-Qualified Domain Name (FQDN) '.
+                             'for the DKIM record instead of the value.');
     }
 
     /**
@@ -43,17 +46,25 @@
      * @param   OutputInterface  $output  An interface to output methods.
      */
     protected function execute(InputInterface $input, OutputInterface $output) {
-      // Print the trimmed record from the database
-      echo trim($this->fetchDKIM(strtolower(trim($input->getArgument('domain'),
-        ".\t\n\r\0\x0B "))))."\n";
+      $domain = strtolower(trim($input->getArgument('domain'),
+        ".\t\n\r\0\x0B "));
+      $info   = $this->fetchDKIMInfo($domain);
+      // Optionally print the FQDN for the DKIM record or ...
+      if ($input->getOption('get-fqdn'))
+        echo trim($info['dkim_selector']).'._domainkey.'.
+             trim($info['name'])."\n";
+      // print the trimmed record from the database
+      else echo trim($info['dkim_record'])."\n";
     }
 
-    protected function fetchDKIM(string $input): ?string {
+    protected function fetchDKIM(string $input): ?array {
       // Query the database for a DKIM record for the requested domain
-      $statement = $this->db->prepare('SELECT `dkim_record` FROM '.
-        '`hosting_schema`.`domains` WHERE `name` = :input LIMIT 0,1');
+      $statement = $this->db->prepare('SELECT `name`, `dkim_record`, '.
+        '`dkim_selector` FROM `hosting_schema`.`domains` WHERE `name` = '.
+        ':input LIMIT 0,1');
       $statement->execute([':input' => $input]);
+      $result = $statement->fetch();
       // Fetch the results from the prepared statement
-      return $statement->fetch()['dkim_record'] ?? null;
+      return is_array($result) ? $result : null;
     }
   }