+++ CHANGELOG.md
@@ -3,6 +3,8 @@
 ## 1.5.4
 
 - Finish all man page documentation.
+- Modify `flush-config`(8) to require `--overwrite` for overwriting PHP-FPM
+  common configuration files from the built-in template.
 
 ## 1.5.3