+++ CHANGELOG.md
@@ -5,6 +5,8 @@
 - Finish all man page documentation.
 - Modify `flush-config`(8) to require `--overwrite` for overwriting PHP-FPM
   common configuration files from the built-in template.
+- Modify `get-dkim`(8) to accept `--get-fqdn` to show the Fully-Qualified Domain
+  Name (FQDN) of a given DKIM record.
 
 ## 1.5.3